Add logic to support string interpolation
To support string interpolation in run, env, if, etc., we will need
the ability to recognize the escape pattern ${{ ... }} and evaluate the
expressions inside.  This adds an initial parser and evaluator for this,
with some documented (as tests) differences from how GitHub Actions
would do this.  We'll need to make some decisions about these cases and
fix or document them accordingly.