Add --verbose to show live command stdout/stderr
If --verbose is specified, then show each command that is being run and
the stdout and stderr from each command as the chunks are received
(on the caller's stdout and stderr respectively).

The verbose command and stdout/stderr are indented so that the normal
test and step names/messages are more clearly indicated. Also, since
"print" uses console.log (implied newline), trim newlines from the end
of each chunk of output/error so that we don't get spurious newlines (we
still could if the chunk boundaries are not at newlines, but it should
be rare).