Add ability to run/exec outer/host commands.
Setting the "exec" key to ":host" will cause the run command to be
spawned from the current process context of dctest (i.e. outside the
docker compose instance).

Refactor docker-exec to make it more generic and split out the
compose-exec part that calls it. Take stdout and stderr streams in the
input and use those (or if not passed in then just create PassThrough
streams. In execute-step* we create streams for stdout and stderr and do
the accumulation and/or verbose echo'ing of the output/error streams.

For the outer exec functionality, this adds an outer-exec function that
is similar to the docker-exec function. outer-exec calls the more
generic outer-spawn which is a promise-based child_process spawn
wrapper.