Support array cmd/args for "run" key.
In the string case, the command(s) will be exec'd using a shell
wrapper. In the array case, the command and args will be exec'd
directly.

The docker exec function already supports this. However, the schema did
not so update the schema to allow the run value to be either a string or
an array. Update the outer-spawn function to support the array style
invocation also.