Warning if no test-suite is specified.
Because of the way that neodoc is parsing the test-suite argument as
"zero or more", it will not error if no test suite is specified. Another
option would be to throw an error here but there might be some
programmatic usage utility in allow the test suite list to be empty as
a no-op (e.g. 0 passes and 0 failures in the results file). But at least
warn (on stderr) when this is the case.