Split interpolation/expr eval. Refactor read/print.
Rename the full text/string interpolation functionality to
`interpolate-text`. The `read-eval-print` function now reads, evaluates,
and prints a string directly containing an expression (not `${{ }}`
wrapped). And the new function `read-eval` does the same but returns the
original evaluated result without printing to a string.

This will enable more flexibility in how and where expressions can be
used.

Detailed changes:
- Rename read-eval-print to interpolate-text.
- Add read-eval and read-eval-print which read an expression (i.e. start
  at the "Expression" grammar node) and return either the evaluated
  result or the printed evaluated result respectively.
- Rename get-ast to read-ast and take an optional starting grammar node
  (default to Expression).
- Add print-ast and print-ast and move the stringifying of evaluated
  resulits out of the eval-ast function and into these printing
  functions.
- Switch tests that just deal with expressions to use read-eval and
  compare to non-stringified values.