Narrow scope of try/catch in execute-step*
Moving the scope of the try/catch to be specifically around the call to
run-exec (either a spawn or exec). We add a :error key with contains
either an unexpected thrown error or an error from an assertion
(currently only the 'exit code == 0' assertion).

This change is not a pure refactoring, as it changes the output to the
results file if --verbose-results is used:

Old --verbose-results for successful and failed steps:

```
{"result":{"code":0,"stdout":"","stderr":""},"outcome":"pass"}
{"error":"Error running command: \"/bin/false\"","outcome":"fail"}
```

New --verbose-results for successful and failed steps:

```
{"code":0,"stdout":"","stderr":"","outcome":"pass"}
{"code":1,"stdout":"","stderr":"","error":{"message":"Error running command: \"/bin/false\""},"outcome":"fail"}
```