Colocate step interpolation logic, use same interface for execs
At a high-level the step logic is:

1) check 'if' condition*, exit early if necessary
2) interpolate step env
3) interpolate rest of step keys using interpolated env
4) run exec on host or container

* 'if' not directly supported yet

In this commit, we are moving steps 1-3 closer to each other and make
new interpolatable keys (e.g. working-directory) straightforward to add.

Additionally, pass context and step directly to outer-exec and
compose-exec to give them a standard input interface (their outputs are
already the same now).