Add name defaults: path for suite, key for test
When a suite does not have an explicit name, use the path as a default.
When a test does not have an explicit name, use the map key as the name.

To enable the default names, tefactor the normalize function to make
suite processing (->suite) similar to to test (->test).