Add 'expect' assertions, support 'step' in expressions
Allow users to add assertions after steps are run using the 'expect'
key, which can be either a string expression or an array of expressions.
We return the error from the first that fails.

Add the results of the step into the context, and allow users to
reference step results using the 'step' context name.