Show values of references within failed expect expressions
When an expression fails, it's helpful to see what the actual values of
all the referenced variables are.  This change walks the AST, returns
all MemberExpression nodes, and returns them as a map with the
evaluated values.

Upgrades nbb for fix to with-out-str and pprint/print-table. Alters the
order of matches in util/indent to properly indent strings with a
newline as the first character.