Allow member access for results of function calls or parens expressions
Previously MemberExpression prohibited something like `foo(env).b.c`,
because `foo(env)` is a FunctionCall, not a ContextName.

We now require that MemberExpression contains at least one Property
access. In order to match `foo(env).b.c` correctly, we must move
MemberExpression to the first matching position of UnaryExpression;
otherwise, FunctionCall will match `foo(env)`, followed by an end of
input error when attempting to match `.b.c`.

Choosing for now to allow member access on Value, but that is up for
debate (ex: `{"a": 3}.a`).