Support method calls
Initial method call support with the first supported methods including
`toString` and `count`.

MemberExpression supports both MethodCall and Property, so that we can
greedily parse expressions like `x.foo().bar.baz().z`. If we attempt to
separate methods and properties into separate expression types, the
MemberExpression will attempt to capture `foo` and `baz` as member
accesses, and we run into end of input errors for the method calls.

MemberExpression evaluation needs to also evaluate its MethodCall
children, so we can pass the referenced object of the MethodCall. This
works but has the drawback of needing to check for uncaught errors again
inside the MemberExpression evaluation.