Catch exceptions thrown in expect expressions
Create a dedicated 'repeat' example that exercises the repeat/retry
functionality, ensuring that we continue retrying if the expect
expression itself throwns an exception.

'explain-refs uses read-eval as well, so we need to catch any exceptions
during that more narrow evaluation of references. (This may be incorrect
behavior. Perhaps we should have drilled through the function call to
show step.stdout only?)