Repeat/re-run exec when 'expect' assertions fail
The 'expect' assertions are conceptually part of the 'run' success; if
'expect' assertions fail, the 'run' should be retried.

We do not need to re-evaluate the 'env' and other keys that are
evaluated before the 'run' executes; however, we do need to re-execute
the 'run' and re-evaluate the 'expect' assertions.