Simplify:  stop returning "mutated" contexts
These test/step execution functions were returning the context for two
reasons:

* they modified the context's :state :failed boolean (which somewhat
  conflates the overall run state and the individual test state)
* they anticipated supporting test and step outputs in the future

However, the :state :failed boolean would then have to be reset to false
in 'run-test, if the user was using --continue-on-error. Additionally,
returning the context meant we had to restore the previous :env.

Instead, it's much simpler to treat the context as immutable! If we need
to support outputs in the future, they should be returned as part of the
test/step result, and the higher level function can determine how to add
those outputs to a newly created context for the next test/step
execution.