Use pending-> to catch and short-circuit on exceptions
Expressions in suite/test/step can throw, in which case we want to stop
evaluating and fail the suite/test/step (ideally preserving as much
information as possible, such as the interpolated :name). This change
adds a pending-> macro to handle the common pattern of:

1. check if the suite/test/step is skipped or failed (if so, short
   circuit)
2. otherwise, attempt to do the next thing (ex: evaluate an env, run a
   test's steps, etc)
3. if an unexpected exception is thrown, catch it, and mark the
   suite/test/step as :fail with an error

Helpers pass!/skip!/fail! are added to manipulate the suite/test/step
outcome consistently and tersely.

The pending-> macro has to be extracted into its own namespace and
duplicated in both clj and cljs files to support ShadowCLJS and nbb.