Show errors from all levels (suite, test, or step)
Now that suites and tests can throw exceptions, we switch to recording
the error on the suite, test, or step as appropriate (previously the
step errors were bubbled up and also assoc'd to the test).

To support this, we summarize errors at all levels in our summary before
displaying the list to the user. As an improvement, we also show where
the error came from (previously the user had to infer which step caused
the failure).