Initial JSON schema for results file
Rename 'results' to 'tests' in the results file, and add a JSON schema
specification for it.  The validateSchema.js script can be used to
validate results files.

Errors are underspecified by the JSON schema, only committing to
providing a message string and allowing any number of other keys that we
do not spec currently.