Add docstrings and comments
Attempt to bring additional clarity to the code by adding docstrings and
comments, making some of the implicit inputs, outputs, assumptions, and
intentions more explicit. Often these are documented in previous commit
messages but not evident in the code itself.

In some cases, adding docstrings or comments may reveal a function is
not yet factored as well as it could be, but the primary goal of this
commit is to mainly document what is there. Any descriptions that feel
excessively long or complicated might can be candidates for future
simplification or refactoring.