Add a guide and reference documentation
Starts a simple prose guide on writing a first test, which references
the other runnable examples as well as new reference documentation.

The reference documentation mirrors the JSON schemas (input and
results-file) and implementation (expressions), but it is unfortunately
not _tied_ to it or linted by it in any way currently. This is an area
for future improvement, but the first goal is getting easy to use, human
readable documentation up.