Render docs with docsify
Docsify allows us to simply commit and push markdown files to create
nicely formatted and navigable documentation with no build step! The
markdown documentation and a short index.html can be served via
GitHub Pages. Docsify fetchs the markdown file via XMLHttpRequest and
renders it as HTML on the fly.

Our docsify configuration points the top level landing page (/) at
the project's root README.md on GitHub to avoid duplication. If
rendering chagnes to the README.md landing page locally, index.html
needs to be temporarily adjusted to use a symlink to the local
README.md (as indicated by the comment).

Additional docsify configuration and plugins are used to use a
non-standard theme and add language syntax highlighting, which is done
on a per-language basis and uses the language annotations from code
fences.

Add docsify dev dependency and npm script for local rendering.