Add --environ-file default for env (using dotenvx).
If the --environ-file option is used then dotenvx is used to load and
parse the environment file to use a defaults env values.

Add an example suite examples/03-env.yaml that is designed to be used
with --environ-file examples/03-env-file.

Add tests to test/runexamples that tests examples/03-env.yaml with and
without the the examples/03-env-file.

NOTE: It would be preferable to have the option be named `--env-file`
since this is more common. However, when running in nbb with node
earlier than 22, node intercepts the --env-file option. And for some
reason, node refused absoluate paths with --env-file. In node v22 it is
possible to avoid the issue by using a "--" parameter before the
regular nbb script parameters. But node 22 is not in standard GHA runner
configs so for now we will use "--environ-file".